IO/Engine構造刷新・Facade再設計・カメラAPI拡張
IOManagerをKeyboard/PadControllerへ刷新しAPIを統一。Engineの各種機能をManager経由に整理し、Game名前空間のFacadeも全面的に再設計。カメラAPIを拡張し切替・情報取得・シェイク制御を強化。FixFPS管理をEngine側に移動。描画登録をDrawSystem経由に変更。HLSL定数バッファ設計を修正しC++側も対応。各所で命名統一・リファクタリング・バグ修正を実施。

+++ DirectXGame/src/engine/DirectX/Pipeline/ShaderReflectionHelper/ShaderReflectionHelper.cpp
@@ -143,7 +143,7 @@ namespace ShaderReflection
                 p.key = "b" + std::to_string(bind.BindPoint);
 
                 p.sizeBytes = cbDesc.Size;
-                p.offsetBytes = currentCBVOffsetBytes;
+                p.offsetBytes = uint32_t(currentCBVOffsetBytes);
 
                 currentCBVOffsetBytes += cbDesc.Size;
                 outParams.push_back(p);
